'3.3.1'

+++ update.groovy
@@ -6,38 +6,46 @@ def process = cmd.execute();
 
 def versions = [];
 process.text.eachLine() {
-        def result = it =~ /v([0-9]*).[0-9]*.[0-9]*/
+    def result = it =~ /v([0-9]*).[0-9]*.[0-9]*/
 
-        def version = result[0][0].startsWith('v')?result[0][0].substring(1):result[0][0];
-        def majorVersion = result[0][1]
+    def version = result[0][0].startsWith('v')?result[0][0].substring(1):result[0][0];
+    def majorVersion = result[0][1]
 
-        if (majorVersion in supportedMajorVersions) {
-            // println "version : $version";
-            // println "majorVersion : $majorVersion";
-            versions << version;
-        }
-
-
+    if (majorVersion in supportedMajorVersions) {
+        // println "version : $version";
+        // println "majorVersion : $majorVersion";
+        versions << version;
+    }
 }
+
 versions = versions.unique()
 
 versions.each() { version->
-    cmd = "mkdir ${version}";
-    String dockerfile = generateDockerfile(version);
 
-    String filePath = "${version}/Dockerfile"
-    def file = new File(filePath)
-    file.newWriter().withWriter { w->
-        w << dockerfile
+    def folder = new File(version);
+    if (!folder.exists()) {
+        "mkdir ${version}".execute().waitFor();
+        String dockerfile = generateDockerfile(version);
+
+        String filePath = "${version}/Dockerfile"
+        def file = new File(filePath)
+        file.newWriter().withWriter { w->
+            w << dockerfile
+        }
+
+        println "cp $filePath ./Dockerfile"
+        "cp $filePath ./Dockerfile".execute().waitFor();
+
+        execute("git add .")
+        execute("git commit -m '${version}'")
+        execute("git tag -f ${version}");
+        execute("git push origin -f ${version}");
+    } else {
+        println "${version} directory exists, skipping";
     }
 
-    println "cp $filePath ./Dockerfile"
-    "cp $filePath ./Dockerfile".execute().waitFor();
 
-    execute("git add .")
-    execute("git commit -m '${version}'")
-    execute("git tag -f ${version}");
-    execute("git push origin -f ${version}");
+
 }
 
 def execute(String cmd) {
